preempt_rt: OCD: make OCD play nice with the p4080 and preempt_rt
CQID:WIND00271588

When CONFIG_WR_OCD_DEBUG is set on the p4080 preempt_rt kernel
the board will fail to boot. The cause was determined to be an
old OCD patch which was required by the PowerQuicc II. Since the
PowerQuicc is no longer a supported BSP we can safely drop this
patch and restore the functionality of the p4080.

The updates to the WRHV patch are to fix the conflict caused
by the dropped patch.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>