preempt-rt: poor performance during SD Card test - new patches
CQID: WIND00281594

1/2 [
Author: Benjamin Herrenschmidt
Email: benh@kernel.crashing.org
Subject: powerpc/book3e: Add generic 64-bit idle powersave support
Date: Wed, 10 Aug 2011 12:08:57 +0800

commit 34d97e07cc81ab6f1e63696127cc7a5d2c4fce4b upstream

We use a similar technique to ppc32: We set a thread local flag
to indicate that we are about to enter or have entered the stop
state, and have fixup code in the async interrupt entry code that
reacts to this flag to make us return to a different location
(sets NIP to LINK in our case).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
[context adjustment]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>
]

2/2 [
Author: Scott Wood
Email: scottwood@freescale.com
Subject: powerpc/e500mc: Add support for the wait instruction in e500_idle
Date: Wed, 10 Aug 2011 12:08:58 +0800

commit ebf714ff37561331eb39963945d80bfc2a59e00f upstream

e500mc cannot doze or nap due to an erratum (as well as having a
different mechanism than previous e500), but it has a "wait" instruction
that is similar to doze.

On 64-bit, due to the soft-irq-disable mechanism, the existing
book3e_idle should be used instead.

Signed-off-by: Vakul Garg <vakul@freescale.com>
Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>
[context adjustment, and add sbcp4080.c fix]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>