mips: tracer: Fix up set_graph_function support for MIPS
CQID: WIND00292624

1/7 [
Author: Wu Zhangjin
Email: wuzhangjin@gmail.com
Subject: MIPS, Tracing: Speed up function graph tracer
Date: Thu, 20 Jan 2011 03:28:27 +0800

[ From mainline commit 9a620a559be65023b5fd5d0eaf37dae884c4f404 ]

This simply moves the "ip-=4" statement down to the end of the do { ...
} while (...); loop, which reduces one unneeded subtration and the
subsequent memory loading and comparison.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>
Cc: Steven Rostedt <srostedt@redhat.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/2006/
Signed-off-by: Ralf Baechle <ralf@duck.linux-mips.net>
]

2/7 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: MIPS, Tracing: Substitute in_kernel_space() for in_module()
Date: Thu, 4 Aug 2011 13:11:46 +0800

[ From mainline commit d9cdb2f1038143c945fcb1a366aae4fa2998419e ]

The old in_module() may not work in some situations(e.g. when module &
kernel are in the same address space when CONFIG_MAPPED_KERNEL=y), The
in_kernel_space() is more generic and it is also easy to be implemented
via cloning the existing core_kernel_text(), so, replace the in_module()
with in_kernel_space().

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>
Cc: Steven Rostedt <srostedt@redhat.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/2005/
Signed-off-by: Ralf Baechle <ralf@duck.linux-mips.net>
]

3/7 [
Author: Wu Zhangjin
Email: wuzhangjin@gmail.com
Subject: MIPS, Tracing: Clean up prepare_ftrace_return()
Date: Thu, 20 Jan 2011 03:28:30 +0800

[ From mainline commit 2816e325969396af5bd1d5f70c7360074ae1d63c ]

The old prepare_ftrace_return() for MIPS is confused and have introduced
some problem. This patch cleans up the names of the arguments, variables
and related functions.

For MIPS, the 2nd argument of prepare_ftrace_return() is not really the
'selfpc' described in ftrace-design.txt but instead it is the self
return address. This did break the compatibility of the generic
interface but really reduced one unneeded calculation for to get the
current function name, the parent return address and the self return
address are enough, no need to tranform the self return address to the
self address.

But set_graph_function of function graph tracer is an exception, it does
need the 2nd argument of prepare_ftrace_return() as 'selfpc', for it
will use 'selfpc' to match user's configuration of function graph
entries, but in reality, it doesn't need the 'selfpc' but the recorded
ip address of the mcount calling site in the __mcount_loc section. So,
the 2nd argument of prepare_ftrace_return() is not important, the real
requirement is the right recorded ip address should be calculated and
assign to trace.func, this will be fixed in the next patches.

Reported-by: Zhiping Zhong <xzhong86@163.com>
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>
Cc: Steven Rostedt <srostedt@redhat.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/2007/
Signed-off-by: Ralf Baechle <ralf@duck.linux-mips.net>
]

4/7 [
Author: Wu Zhangjin
Email: wuzhangjin@gmail.com
Subject: MIPS, Tracing: Clean up ftrace_make_nop()
Date: Thu, 20 Jan 2011 03:28:31 +0800

[ From mainline commit 7f21a60968221eabad5c53fe760db3d094994011 ]

This moves the comments out of ftrace_make_nop() and cleans it.  At the
same time, a macro MCOUNT_OFFSET_INSNS is defined for sharing with the
next patch.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>
Cc: Steven Rostedt <srostedt@redhat.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/2008/
Signed-off-by: Ralf Baechle <ralf@duck.linux-mips.net>
]

5/7 [
Author: Wu Zhangjin
Email: wuzhangjin@gmail.com
Subject: MIPS, Tracing: Fix set_graph_function of function graph tracer
Date: Sat, 22 Jan 2011 02:01:53 +0800

[ From mainline commit b9f07eb2f25a64098e2ba223c1a2fe2a8f249e01 ]

trace.func should be set to the recorded ip of the mcount calling site
in the __mcount_loc section to filter the function entries configured
through the tracing/set_graph_function interface, but before, this is
set to the self_ra(the return address of mcount), which has made
set_graph_function not work as expected.

This fixes it via calculating the right recorded ip in the __mcount_loc
section and assign it to trace.func.

Reported-by: Zhiping Zhong <xzhong86@163.com>
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>
Cc: Steven Rostedt <srostedt@redhat.com>
Cc: Sergei Shtylyov <sshtylyov@mvista.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/2017/
Signed-off-by: Ralf Baechle <ralf@duck.linux-mips.net>
]

6/7 [
Author: David Daney
Email: ddaney@caviumnetworks.com
Subject: MIPS: Mask jump target in ftrace_dyn_arch_init_insns().
Date: Tue, 28 Dec 2010 13:21:37 -0800

[ From mainline commit c54794d19e61472156e37263c074225574c80df1 ]

The current code is abusing the uasm interface by passing jump target
addresses with high bits set.  Mask the addresses to avoid annoying
messages at boot time.

Signed-off-by: David Daney <ddaney@caviumnetworks.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Wu Zhangjin <wuzhangjin@gmail.com>
Patchwork: https://patchwork.linux-mips.org/patch/1922/
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
]

7/7 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: ftrace/graph: MIPS: Trace function entry before updating index
Date: Thu, 11 Aug 2011 14:59:20 +0800

[ Based on mainline commit 722b3c74695377d11d18a52f3da08114d37f3f37 ]

Currently the index to the ret_stack is updated and the real return address
is saved in the ret_stack. Then we call the trace function. The trace
function could decide that it doesn't want to trace this function
(ex. set_graph_function does not match) and it will return 0 which means
not to trace this call.

The normal function graph tracer has this code:

if (!(trace->depth || ftrace_graph_addr(trace->func)) ||
      ftrace_graph_ignore_irqs())
	return 0;

What this states is, if the trace depth (which is curr_ret_stack)
is zero (top of nested functions) then test if we want to trace this
function. If this function is not to be traced, then return  0 and
the rest of the function graph tracer logic will not trace this function.

The problem arises when an interrupt comes in after we updated the
curr_ret_stack. The next function that gets called will have a trace->depth
of 1. Which fools this trace code into thinking that we are in a nested
function, and that we should trace. This causes interrupts to be traced
when they should not be.

The solution is to trace the function first and then update the ret_stack.

Reported-by: zhiping zhong <xzhong86@163.com>
Reported-by: wu zhangjin <wuzhangjin@gmail.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
[ Apply for MIPS ]
Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>