ti_sitara_am389x: add kernel template
This is a new Async BSP merging to kernel-4.3, it supports standard
and preempt_rt ktypes.

This BSP has 43 kernel patches for $BSP-standard branch, after this
merging, 41 of them are moved to -std branch, 2 of them are left to
$BSP-standard branch. Besides these 43 patches, it also has 14
preempt_rt specific kernel patches, 1 of them is common and is moved
to preempt_rt branch, 13 of them are left to $BSP-preempt_rt branch.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>