meta/qemuppc: board description fixes
The qemuppc rt board description did not match the recent change to
a board specific branch on the end of the preempt-rt kernel type. This
resulted in kernel configuration validation errors since the board
description didn't match the tree and was discarded. This updates the
.scc file to have a proper branch and series description.

The .cfg file is also update to remove obselete config values and provide
a clean configuration audit.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>