Feature Testing: Networking: On cn3860, 2 test cases fail in TAHI Self_
CQID: WIND00269107

1/1 [
Author: Roy.Li
Email: rongqing.li@windriver.com
Subject: Software should check the checksum of no tcp/udp packets
Date: Thu, 8 Sep 2011 15:21:18 +0800

When icmpv6_rcv handles the packets, it would not verify the icmp
checksum if the skb->ip_summed is CHECKSUM_UNNECESSARY.

The L4_error register can report L4 checksum error, if it is Zero,
it means the L4 checksum verification is OK, and does not need
software to verify the checksum again, so CHECKSUM_UNNECESSARY is
set.

But L4_error only works for TCP/UDP, not for ICMP from test result.

Signed-off-by: Roy.Li <rongqing.li@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>