arm_coretile_express_a9x4: fix ftrace self test - standard branch
CQID: WIND00302757

1/5 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: sched_clock: provide common infrastructure for sched_clock()
Date: Wed, 15 Dec 2010 19:23:07 +0000

commit 112f38a4a31668eb6a7d91d128296a26afdf7c4b upstream

Provide common sched_clock() infrastructure for platforms to use to
create a 64-bit ns based sched_clock() implementation from a counter
running at a non-variable clock rate.

This implementation is based upon maintaining an epoch for the counter
and an epoch for the nanosecond time.  When we desire a sched_clock()
time, we calculate the number of counter ticks since the last epoch
update, convert this to nanoseconds and add to the epoch nanoseconds.

We regularly refresh these epochs within the counter wrap interval.
We perform a similar calculation as above, and store the new epochs.

We read and write the epochs in such a way that sched_clock() can easily
(and locklessly) detect when an update is in progress, and repeat the
loading of these constants when they're known not to be stable.  The
one caveat is that sched_clock() is not called in the middle of an
update.  We achieve that by disabling IRQs.

Finally, if the clock rate is known at compile time, the counter to ns
conversion factors can be specified, allowing sched_clock() to be tightly
optimized.  We ensure that these factors are correct by providing an
initialization function which performs a run-time check.

Acked-by: Peter Zijlstra <peterz@infradead.org>
Tested-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Tested-by: Will Deacon <will.deacon@arm.com>
Tested-by: Mikael Pettersson <mikpe@it.uu.se>
Tested-by: Eric Miao <eric.y.miao@gmail.com>
Tested-by: Olof Johansson <olof@lixom.net>
Tested-by: Jamie Iles <jamie@jamieiles.com>
Reviewed-by: Nicolas Pitre <nicolas.pitre@linaro.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
]

2/5 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: sched_clock: allow init_sched_clock() to be called early
Date: Tue, 11 Jan 2011 16:23:04 +0000

commit 211baa7016894c02fc18693e21ca479cd08ac0c0 upstream

sched_clock is supposed to be initialized early - in the recently added
init_early platform hook.  However, in doing so we end up calling
mod_timer() before the timer lists are initialized, resulting in an
oops.

Split the initialization in two - the part which the platform calls
early which starts things off.  The addition of the timer can be
delayed until after we have more of the kernel initialized - when the
normal time sources are initialized.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[minor change to apply the patch]
Integrated-by: Junxiao Bi <junxiao.bi@windriver.com>
]

3/5 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: sched_clock: make minsec argument to clocks_calc_mult_shift() zero
Date: Tue, 11 Jan 2011 16:44:02 +0000

commit edc4d272551594729f63ca4cde1612608494091f upstream

The purpose of the minsec argument is to prevent 64-bit math overflow
when the number of cycles is multiplied up.  However, the multipler
is 32-bit, and in the sched_clock() case, the cycle counter is up to
32-bit as well.  So the math can never overflow.

With a value of 60, and clock rates greater than 71MHz, the calculated
multiplier is unnecessarily reduced in value, which reduces accuracy by
maybe 70ppt.  It's almost not worth bothering with as the oscillator
driving the counter won't be any more than 1ppm - unless you're using
a rubidium lamp or caesium fountain frequency standard.

So, set the minsec argument to zero.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
]

4/5 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: Realview/Versatile: separate out common sched_clock()
Date: Thu, 14 Jan 2010 20:07:35 +0000

commit f5fc00826d1e60af0e22cb9f65b933d823a8ed84 upstream

Provide a common sched_clock() implementation for Versatile and
Realview.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[remove the changes to mach-versatile and mach-realview since they haven't switch
to plat-versatile in wrlinux]
Signed-off-by: Junxiao Bi <junxiao.bi@windriver.com>
]

5/5 [
Author: Russell King
Email: rmk+kernel@arm.linux.org.uk
Subject: ARM: versatile: convert sched_clock() to use new infrastructure
Date: Wed, 15 Dec 2010 21:56:47 +0000

commit 1da0c89c66753860ccfe81eb327c25db46c2a24a upstream

Convert versatile platforms to use the new sched_clock() infrastructure
for extending 32bit counters to full 64-bit nanoseconds.

Tested-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[remove the changes to mach-versatile and mach-realview since they haven't switch
to plat-versatile in wrlinux]
Signed-off-by: Junxiao Bi <junxiao.bi@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>