ARM GOS - Avoid GIC if no direct irq
CQID: WIND00303219

1/1 [
Author: Jim Somerville
Email: Jim.Somerville@windriver.com
Subject: wrhv: arm: Avoid GIC accesses if no direct interrupts
Date: Wed, 14 Sep 2011 12:04:20 -0400

If the user does not specify direct_interrupts= on
the BootLine, then direct interrupts are not to be
used so the GIC should not be touched at all.

This change is to facilitate core sharing between
VBs.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>