mips/nlm_xlp_64_be: enable NAE in the 2nd kernel booted by kexec
CQID: WIND00289653

1/1 [
Author: Jin Yanjiang
Email: yanjiang.jin@windriver.com
Subject: mips/nlm_xlp_64_be: enable NAE in the 2nd kernel booted by kexec
Date: Wed, 14 Sep 2011 15:57:44 +0800

If you boot a 2nd kernel via kexec it will hang if the frequency is
set again. To fix this issue, we should not call the function
set_nae_frequency() on booting in the 2nd kernel.
The function set_nae_frequency() will set the net field of
SYS_DFS_DIV_DEC_CTRL to 1 for NAE.
As the Programming Reference Manual of XLP Processor said:
"Setting a read/write field to 1 will increment the DFS clock divider for
a device or interface by one."
The system clock of net was set by the first kernel, and was not cleared
after the secondary kernel booted up.
So if the secondary kernel set it again, the NAE frequency will double, and
then system will hang.

Signed-off-by: Jin Yanjiang <yanjiang.jin@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>