nfsroot: Add mport to allow specification of the mnt_port on an NFS root mount
CQID: WIND00309125

1/1 [
Author: Jason Wessel
Email: jason.wessel@windriver.com
Subject: nfsroot: Add mport to allow specification of the mnt_port on an NFS root mount
Date: Mon, 3 Oct 2011 14:57:14 -0500

It is possible to entirely avoid contacting rpcbind/portmap
services if assigned ports are in use for the NFS root mount.

The main user of this functionality is booting up a kernel
and connecting it to a user mode NFS server on a host
where rpcbind/portmap services are not running.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>