cgl: fix up bogosity in NUMA slub patch
CQID: 243493

The cherry pick of upstream 4a5fa3590f09999f6db41bc386bce40848fa9f63 for
stable clearly shows that slub requires NUMA, so this cgl patch putting a
depends on !NUMA on slub must be wrong.  Based on the description in the
patch, they wanted to exclude NUMA && GRSEC && SLUB, so make it do just
that, and leave the non GRSEC folks free to do whatever they want.

This also fixes the conflict with the above stable commit.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>