net/gianfar: workaround for incomplete GRS when 1588 is enabled
CQID: WIND00310770

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: net/gianfar: workaround for incomplete GRS when 1588 is enabled
Date: Tue, 11 Oct 2011 13:14:58 +0800

On p2020 when 1588 timer is enabled, the GRS can't complete.
This scenario is very similar to errata eTSEC-A002. The workaround
for eTSEC-A002 has used another method to check whether the Rx is
idle. Also use this method to resolve this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>