tracing: Replace syscall_meta_data struct array with pointer array
CQID: WIND00307849

1/1 [
Author: Liming Wang
Email: liming.wang@windriver.com
Subject: tracing: Replace syscall_meta_data struct array with pointer array
Date: Sat, 8 Oct 2011 16:32:04 +0800

[ commit 3d56e331b6537671c66f1b510bed0f1e0331dfc8 upstream
 backport from mainline and modify a little ]

Currently the syscall_meta structures for the syscall tracepoints are
placed in the __syscall_metadata section, and at link time, the linker
makes one large array of all these syscall metadata structures. On boot
up, this array is read (much like the initcall sections) and the syscall
data is processed.

The problem is that there is no guarantee that gcc will place complex
structures nicely together in an array format. Two structures in the
same file may be placed awkwardly, because gcc has no clue that they
are suppose to be in an array.

A hack was used previous to force the alignment to 4, to pack the
structures together. But this caused alignment issues with other
architectures (sparc).

Instead of packing the structures into an array, the structures' addresses
are now put into the __syscall_metadata section. As pointers are always the
natural alignment, gcc should always pack them tightly together
(otherwise initcall, extable, etc would also fail).

By having the pointers to the structures in the section, we can still
iterate the trace_events without causing unnecessary alignment problems
with other architectures, or depending on the current behaviour of
gcc that will likely change in the future just to tick us kernel developers
off a little more.

The __syscall_metadata section is also moved into the .init.data section
as it is now only needed at boot up.

Suggested-by: David Miller <davem@davemloft.net>
Acked-by: David S. Miller <davem@davemloft.net>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Integrated-by: Liming Wang <liming.wang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>