powerpc: ocd: put back setting of MSR DE bit
CQID: WIND00313314

In commit 85aae119 a patch was removed from the kernel which
was specific to the powerquicc, a BSP we no longer support.
Some of this patch, however, should have been left otherwise
the MSR DE bit is not properly set during the initial boot.
Not having the MSR DE bit on during boot would cause issues
if you attempted to use debug features during the boot sequence.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>