intel_xeon_5520_ioh/kernel-cache: enable numa in the kernel config
For this BSP, we need numa to be enabled by default, we added a
"CONFIG_NUMA=y" in the config fragment, but in the config fragment,
there is already an item to disable numa like this
"# CONFIG_NUMA is not set".

So to make "CONFIG_NUMA=y" take effect, we need to remove
"# CONFIG_NUMA is not set".

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>