common_pc/kernel_cache: enable FB_CIRRUS by default
If we want to run some framebuffer dependent programs like directfb,
we should enable framebuffer driver. For common_pc BSP, we have
already enabled drm_i915 and KMS, this can bring framebuffer driver
work well on the real platform (ASUS Eee PC), but in the qemu, it
doesn't, we need to enable FB_CIRRUS for the qemu.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>