fix KFTP testcase mthread-lock-on-futex call trace and board hang
CQID: WIND00315585

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc/trace: add a dummy stack frame for trace_hardirqs_off
Date: Tue, 8 Nov 2011 16:33:37 +0800

The trace_hardirqs_off will use CALLER_ADDR0 and CALLER_ADDR1.
If an exception occurs in user mode, there is only one stack frame
on the stack and accessing the CALLER_ADDR1 will cause oops. So we
create a dummy stack frame to make trace_hardirqs_off happy.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc: don't enable irq in program exception
Date: Wed, 9 Nov 2011 22:32:50 +0800

The commit cd8a5673 (powerpc: Fix might-sleep warning in program
check exception handler) caused a subtle regression in the irq in
program exception handler in the kprobe return from interrupt in that
it enables interrupts when they should not have.  There is a dedicated
stack for program exception when kprobes is enabled, and this requires
that the program exception handler must be nonpreemptive.
The instruction that is being read was actually the faulting
instruction so there should not be any further data faults.
In this case we can replace get_user() with probe_kernel_read()
and keep the IRQs disabled.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Reviewed-by: Jason Wessel <jason.wessel@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>