for standard, patches/net - sustaining pull 11-11-15
CQID: WIND00308423

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: sctp: fix potential reference of a freed pointer
Date: Sun, 9 Oct 2011 09:20:09 +0800

commit 0c42749cffbb4a06be86c5e5db6c7ebad548781f upstream

sctp: fix potential reference of a freed pointer

When sctp attempts to update an assocition, it removes any
addresses that were not in the updated INITs.  However, the loop
may attempt to refrence a transport with address after removing it.

Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Integrated-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>