Add EFI scc and cfg files
Basic EFI support only requires CONFIG_EFI=y, this is sufficient for
some boards, and desirable for small configs. This is done with efi.scc.

Additional support for CONFIG_EFI_VARS, CONFIG_EFI_PARTITION, and CONFIG_FB_EFI
is provided via efi-ext.scc (extended) as this pulls in the block layer,
framebuffer support, and virtual terminals.

I'd like EFI_VARS to be part of the base config, but I have received
reports of it failing in some situations. Keeping it separate ensures
basic boot can work with the fragments as defined.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>