broadcom_bcm911107evm: bcmring_net fix different head between capture and really send
CQID: WIND00284246

1/1 [
Author: jiaxi liu
Email: jiaxi.liu@windriver.com
Subject: bcmring_net fix different head between capture and really send
Date: Wed, 23 Nov 2011 15:41:48 +0800

We observe this issue on Broadcom network device driver:
1. ping command could work fine, no lost packet.
2. use tcpdump or custom test case could discover that the packet's
head is different between capture and really send.

Here is some output showing the first 32 octets of capture buffer
   8e 6b b4 b0 17 81 a2 86  1c 00 00 00 08 00 45 00
   03 dd ae 48 40 00 40 06  04 86 c0 a8 01 99 c0 a8

In that case, the content should be
   destination mac    00:24:8c:03:8e:6b
   source mac         b4:b0:17:81:a2:86

but in fact, the destination mac is missing the first 4 bytes.

Cause:
When packets are processed, bytes 0-16 in the packet header are
re-written by the driver with a Broadcom specific header that conveys
management instructions to the management block running on the CPU.

The re-write of the header occurs after capture and before the send,
which makes the captured packet and sent packet different.

Fix this issue by make a private copy before sending the data.

Signed-off-by: jiaxi liu <jiaxi.liu@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>