mmc: sdhci-esdhc: fix kmmcd high cpu usage
CQID: WIND00328010

1/1 [
Author: Junxiao Bi
Email: junxiao.bi@windriver.com
Subject: mmc: sdhci-esdhc: fix kmmcd high cpu usage
Date: Fri, 20 Jan 2012 15:23:17 +0800

esdhc_set_clock() can be invoked in a context when a spin lock is holded,
as this function is invoked every 1 second, mdelay(100) will cause high
cpu usage. This delay is to wait hardware clock working. So we move it
out of the sleep-not-allowed context and use msleep() to replace it.
For mmc_go_idle(), it is invoked in a sleep-allowed context, using msleep
to replace mmc_delay can also lower the cpu usage.

Signed-off-by: Junxiao Bi <junxiao.bi@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>