fri2: Drop 8250 from the fri2 scc
The FRI2 has a pch_uart. Depending on the firmware, the kernel may
decide to use the 8250 or the pch_uart driver. As the firmware
configures the pch_uart with a non-standard UART clock, it is
necessary to ensure the pch_uart is used. To do this, drop the 8250
support from the kernel.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>