TIPC: include upstream commits that close out sourceforge backlog
There was a queue of fixes based on tipc-1.7 that were stacked up
in sourceforge but not merged upstream.  This queue was finally
closed out during the 3.5 kernel development cycle.

As this last batch was tested on a 3.4 baseline, and since it is
a standalone subsystem, we can easily stack them on for 3.4 usage
here as well.

The 3.5 content also had a large 4000 line patch that compressed
out gratuitous line feeds, but that has been omitted here.

Note that standard-nocfg.scc already hooks in features/tipc.scc so
such an addition is not present or required in this commit.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>