cfg: delete instances of any reference to dmesg buf size.
Due to historical reasons, the standard.cfg and rt.cfg were
setting a LOG_BUF_SHIFT value that was (at that point in time)
larger than the mainline kernel default value.  But time has
moved on, and now it is smaller.

This in turn spawned a features/logbuf/size-normal.cfg fragment
that set a larger value, but time again has moved past this too,
and so the "size-normal" value is once again smaller than the
mainline default (which is CONFIG_LOG_BUF_SHIFT=18).

Get rid of all references; either directly or via reference to
the size-normal thing, and just implicitly take the mainline
default, so we grow with it and don't have to revisit this
in another year or two.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>