e500mc-kvm-guest: fix external proxy register
CQID: WIND00336196

The e500mc have an extra external proxy register which is not supported
in kvm. In order to boot e500mc with interrupts enable, the guest need to
use the normal mpic_get_irq(). Otherwise, once an interrupt comes in, the
guest will not be able to acknowledge the interrupt and hang.

Signed-off-by: Thomas Tai <thomas.tai@windriver.com>