x86: pvclock: Add flag to indicate that a vm was stopped by the host
CQID: WIND00335963

1/1 [
Author: Eric B Munson
Email: emunson@mgebm.net
Subject: x86: pvclock: Add flag to indicate that a vm was stopped by the host
Date: Thu, 28 Jun 2012 12:51:12 -0700

This flag will be used to check if the vm was stopped by the host when a soft
lockup was detected.  The host will set the flag when it stops the guest.  On
resume, the guest will check this flag if a soft lockup is detected and skip
issuing the warning.

Signed-off-by: Eric B Munson <emunson@mgebm.net>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>